requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.3
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
efficientnet==1.0.0
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==25.1.24
focal-loss==0.0.7
fonttools==4.55.8
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h5py==3.12.1
idna==3.10
image-classifiers==1.0.0
imageio==2.37.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
keras==3.8.0
keras-applications==1.0.8
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.6
pydantic-core==2.27.2
pydot==3.0.4
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
segmentation-models==1.0.1
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
stack-data==0.6.3
stringzilla==3.11.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2

flask
tensorflow
pillow
numpy
gunicorn